Modify wxGUI as a result of credit mining refactor
